+++ sh_sms/models/res_partner.py
@@ -3,42 +3,103 @@
 
 from odoo import api, fields, models, _
 from datetime import date,datetime,time,timedelta
-# import pytz
+from twilio.rest import Client
+from twilio.base.exceptions import TwilioException
 
 class ResPartner(models.Model):
     _inherit = 'res.partner'
 
     sh_dob=fields.Date(string='Date of Birth')
-    sh_send_sms_count = fields.Integer(string='Sent SMS Count',compute='_compute_action_send_sms_count')
+    sh_send_sms_count = fields.Integer(string='Sent SMS Count')
     sh_customer_bool = fields.Boolean(string='SMS Customer',default=False)
-    
-    def _compute_action_send_sms_count(self):
-        for rec in self:
-            contact_sms_count = self.env['sh.sms.history'].search_count([('sh_partner_id','=',rec.id)])
-
-            rec.sh_send_sms_count = contact_sms_count
-
+    sh_top_store = fields.Many2one('sh.client.shop',string='Top Store')
+   
     # WISHING TODAY BIRTDAY SMS SEND METHOD 
+    # CRON METHOD 
     @api.model
     def wish_birthday(self):
-        today = date.today()
-        print('\n\n\n <----------Today Date-------->',today)
-        today = date.today()
-        partner_id=self.env['res.partner'].search([('sh_dob','=',today)])
-        print('\n\n\n\n <-------partner-------->',partner_id)
+
+        customer = self.env['res.partner'].search([('sh_dob','=',fields.Date.today())])
+        birthday_sms_template = self.env.company.sh_birthday_sms_temp_id
+        twilio_account = self.env['sh.twilio.account'].search([('state','=','confirm')],limit=1)
+        
+        if not birthday_sms_template:
+            #TODO: FIALED SEND SMS DEUT TO MOBILE NUMBER IS NOT SET IN CUSTOMER
+            # for rec in customer:
+                self.env['sh.sms.history'].create({
+                    'sh_message':'Started birthday wishing reminder cron: Birthday SMS template is not defined.',
+                    'sh_state':'fail'
+                })
+                return 
+
+        if not twilio_account:
+            #TODO: FIALED SEND SMS DEUT TO MOBILE NUMBER IS NOT SET IN CUSTOMER
+            # for rec in customer:
+                self.env['sh.sms.history'].create({
+                    'sh_message':'Started birthday wishing reminder cron: The twilio account details invalid.',
+                    'sh_state':'fail'
+                })
+                return 
+      
+        if customer and birthday_sms_template.body and twilio_account:
+            client = Client(twilio_account.sh_account_sid,
+                twilio_account.sh_auth_token)
+                   
+            for rec in customer:
+                try :
+                    if rec.mobile:
+                        
+                        # RENDER TEMPLATE 
+                        body = birthday_sms_template._render_template(birthday_sms_template.body,'res.partner',rec.ids)
+
+                        message = client.messages.create(
+                            body= body.get(rec.id),
+                            from_=twilio_account.sh_from_number,
+                            to=rec.mobile
+                        )
+
+                        self.env['sh.sms.history'].create({
+                            'sh_partner_id':rec.id,
+                            'sh_store_id':rec.sh_top_store.id,
+                            'sh_message':body.get(rec.id),
+                            'sh_state':'sent'
+                        })
+                        rec.sh_send_sms_count+=1
+
+                    if not rec.mobile:
+                        #TODO: FIALED SEND SMS DEUT TO MOBILE NUMBER IS NOT SET IN CUSTOMER
+                        self.env['sh.sms.history'].create({
+                            'sh_partner_id':rec.id,
+                            'sh_store_id':rec.sh_top_store.id,
+                            'sh_message':'Started birthday wishing reminder cron: The mobile number is empty.',
+                            'sh_state':'fail'
+                        })
+
+                except TwilioException as e:
+                    #TODO: FIALED SEND SMS DEUT TO MOBILE NUMBER IS NOT SET IN CUSTOMER
+                    self.env['sh.sms.history'].create({
+                            'sh_partner_id':rec.id,
+                            'sh_store_id':rec.sh_top_store.id,
+                            'sh_message':f'Started birthday wishing reminder cron: Twilio account details invalid ..! {e}',
+                            'sh_state':'fail'
+                        })
 
     # TIME PERIOD UNDER NOT SALE_ORDER CREATE THAT PARTNER SMS SEND METHOD  
+    # CRON METHOD 
+
     @api.model
     def store_visit_customer_sms(self):
-        day = self.env['res.company'].browse(self.env.company.id).sh_sms_day
-        print('\n\n\n\n <----Day---->',day)
+        day = self.env.company.sh_sms_day
+        # print('\n\n\n\n DAY',day)
+
+        store_person_sms_template = self.env.company.sh_store_visit_sms_temp_id
+        # print('\n\n\n\n STORE_PERSON_SMS',store_person_sms_template)
+
+        twilio_account = self.env['sh.twilio.account'].search([('state','=','confirm')],limit=1)
 
         # if day >= 1:
-
         previous_date = date.today() - timedelta(day)
-        print('\n\n\n\n <-----previous_date------>',previous_date)
-        # new_date_time_object = datetime.combine(previous_date, time(00,00,00))
-        # print('\n\n\n <--------previous_date----->',new_date_time_object)
+        # print('\n\n\n\n <-----previous_date------>',previous_date)
 
         query ="""
             SELECT partner_id
@@ -52,22 +113,73 @@ class ResPartner(models.Model):
         self._cr.execute(query, parameter)    
 
         result = self._cr.dictfetchall()
-        print('\n\n\n\n <----------result------------>',result)
 
+        partner =[]        
         if result:
-            partner= [rec.get('partner_id') for rec in result]
-            print('\n\n\n <--RESULT----->',partner)
+            partner.extend(rec.get('partner_id') for rec in result)
+            # print('\n\n\n\n PARTNER',partner)
+
+            if partner and not twilio_account:
+                # customer_send_msg = self.env["res.partner"].sudo().search([('id','not in',partner),('sh_customer_bool','=',False)])
+
+                self.env['sh.sms.history'].create({
+                    'sh_message':'Started Store vist reminder cron: The twilio account details invalid.',
+                    'sh_state':'fail'
+                })
+                return
+
+            if partner and not store_person_sms_template:
+                self.env['sh.sms.history'].create({
+                    'sh_message':'Started store visit reminder cron: Store visit template not set in settings.',
+                    'sh_state':'fail'
+                })
             
-            if partner:
+                return
+
+            if partner and twilio_account:
                 customer_send_msg = self.env["res.partner"].sudo().search([('id','not in',partner),('sh_customer_bool','=',False)])
-                print('\n\n\n\n <-------customer_send_msg------->',customer_send_msg)
+                client = Client(twilio_account.sh_account_sid,twilio_account.sh_auth_token)
 
-                if customer_send_msg:
-                    customer_bool_update = [rec.write({'sh_customer_bool':True}) for rec in customer_send_msg]
+                for rec in customer_send_msg:
+                    try:
+                        if rec.mobile:
+                            body = store_person_sms_template._render_template(store_person_sms_template.body,'res.partner',rec.ids)
 
+                            message = client.messages.create(
+                                body= body.get(rec.id),
+                                from_=twilio_account.sh_from_number,
+                                to=rec.mobile
+                            )
+                            
+                            self.env['sh.sms.history'].create({
+                                'sh_partner_id':rec.id,
+                                'sh_store_id':rec.sh_top_store.id,
+                                'sh_message':body.get(rec.id),
+                                'sh_state':'sent'
+                            })
+                            rec.sh_send_sms_count+= 1
+                            rec.sh_customer_bool= True
+                        
+                        if not rec.mobile:
+                            self.env['sh.sms.history'].create({
+                                'sh_partner_id':rec.id,
+                                'sh_store_id':rec.sh_top_store.id,
+                                'sh_message':'Started store visit reminder cron: The mobile number is empty.',
+                                'sh_state':'fail'
+                            })
+
+                    except TwilioException as e:
+                        #TODO: FIALED SEND SMS DEUT TO MOBILE NUMBER IS NOT SET IN CUSTOMER
+                        self.env['sh.sms.history'].create({
+                            'sh_partner_id':rec.id,
+                            'sh_store_id':rec.sh_top_store.id,
+                            'sh_message':f'Started store visit reminder cron: Twilio account details invalid ..! {e}',
+                            'sh_state':'fail'
+                        })
+    
     # IR.ACTION.SERVER METHOD
     def action_sms_text_message(self):
-        action = {
+        return {
             'name': _(' Send SMS Text Message'),
             'res_model': 'sh.sms.text.message',
             'type': 'ir.actions.act_window',
@@ -75,12 +187,10 @@ class ResPartner(models.Model):
             'target':'new',
         }
 
-        return action
-
     # SMS HISTORY METHOD 
     def sh_sms_history(self):
         
-        action= {
+        return {
             'name': _(' SMS History'),
             'res_model': 'sh.sms.history',
             'type': 'ir.actions.act_window',
@@ -90,29 +200,22 @@ class ResPartner(models.Model):
             # 'domain':"[('sh_partner_id','=',context.get('default_partner_id'))]",
             'domain':[('sh_partner_id','=',self.id)]
         }  
-        print(action)
-
-        return action  
 
     # SEND SMS METHOD 
     def sh_send_sms(self):
-        print('\n\n\n self',self)
-
-        action = {
+        # print('\n\n\n self',self)
+        return {
             'name': _(' Send SMS Text Message'),
             'res_model': 'sh.sms.text.message',
             'type': 'ir.actions.act_window',
             'view_mode': 'form',
             'target':'new',
         }
-    
-        return action
 
     # Details Loyalty Point METHOD
     def sh_details_loyalty_points(self):
-        print('\n\n\n self',self)
-
-        action = {
+        # print('\n\n\n self',self)
+        return {
             'name': _(' Details Loyalty Point'),
             'res_model': 'sh.loyalty.point',
             'type': 'ir.actions.act_window',
@@ -120,4 +223,3 @@ class ResPartner(models.Model):
             'target':'new',
         }
     
-        return action