+++ sh_po_customisation/doc/changelog.rst
@@ -5,3 +5,8 @@
 16.0.2(Date: 06 June 2024)
 ----------------------------------
 - Purchase Order Under Unit Price Field digits parameter update.
+
+16.0.3(Date: 26 June 2024)
+----------------------------------
+- Purchase_Order_Line Product related Price_Unit & Barcode Field Update,
+  so That Product related Stnadard_Price & Barcode field updated.